requirements.txt
torch
torchvision
torchaudio
ultralytics
pillow
pandas
numpy
yolov5